requirements.txt
Flask==2.0.1
python-dotenv==1.0.1
Werkzeug==3.0.1
torch==2.0.0
torchaudio==2.0.0
openai-whisper==20231117
numpy==1.26.4
tqdm==4.66.2
ffmpeg-python==0.2.0
python-docx==0.8.11
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
python-jose==3.3.0
passlib==1.7.4
python-magic==0.4.27
nvidia-cudnn-cu11==8.9.4.25
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cublas-cu11==11.11.3.6
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
nvidia-nvtx-cu11==11.8.86 